complete original vs refactor mps test (.py)
