+++ python_prototype/outputs/lds_sequences/magic_point_shop_times.csv
@@ -1,0 +1,21 @@
+n_2power,mps_lattice_time,qmcpy_lattice_time,mps_Sobol_time,qmcpy_Sobol_time
+1.0,5.8650970458984375e-05,3.62396240234375e-05,0.8803889751434326,0.014493465423583984
+2.0,6.699562072753906e-05,4.220008850097656e-05,0.779200553894043,0.012497425079345703
+3.0,0.00011396408081054688,7.867813110351562e-05,0.8854105472564697,0.013801336288452148
+4.0,0.00013947486877441406,0.00011110305786132812,0.8279976844787598,0.01523733139038086
+5.0,0.0002429485321044922,0.00033664703369140625,0.8117339611053467,0.012526512145996094
+6.0,0.00040030479431152344,0.0003571510314941406,0.7708349227905273,0.015253305435180664
+7.0,0.0009064674377441406,0.0008454322814941406,0.7857370376586914,0.012649774551391602
+8.0,0.0019223690032958984,0.0014617443084716797,0.7563800811767578,0.01493215560913086
+9.0,0.0032384395599365234,0.0030612945556640625,0.8273439407348633,0.017081499099731445
+10.0,0.0074138641357421875,0.005825996398925781,0.8533236980438232,0.017297744750976562
+11.0,0.014200687408447266,0.013368844985961914,0.8003942966461182,0.017540931701660156
+12.0,0.046332597732543945,0.023044586181640625,0.7951755523681641,0.022562265396118164
+13.0,0.05243492126464844,0.046317100524902344,0.8220031261444092,0.03580117225646973
+14.0,0.10132956504821777,0.09380149841308594,0.838721752166748,0.05526542663574219
+15.0,0.21720290184020996,0.18790555000305176,0.9474959373474121,0.09707880020141602
+16.0,0.39559340476989746,0.3976893424987793,1.093688726425171,0.1843724250793457
+17.0,0.8527541160583496,0.779766321182251,1.4314684867858887,0.35771727561950684
+18.0,1.6234207153320312,1.5915172100067139,2.1369638442993164,0.6914055347442627
+19.0,3.3655686378479004,3.244432210922241,3.435457706451416,1.4022958278656006
+20.0,6.802563667297363,6.553140878677368,6.100977659225464,2.794978618621826
