+++ python_prototype/outputs/mc_vs_qmc/abstols_plot.png
