+++ python_prototype/outputs/mc_vs_qmc/dimension_plot.png
